Code correction

+++ BioradMedisys.DataAccessLayer/DataEntity/UserMaster.cs
@@ -24,9 +24,9 @@ namespace Coditech.DataAccessLayer.DataEntity
         public bool IsActive { get; set; }
         public Nullable<byte> AdminRoleMasterId { get; set; }
         public Nullable<bool> IsDocumentApprovalAuthority { get; set; }
-        public Nullable<long> CreatedBy { get; set; }
+        public Nullable<int> CreatedBy { get; set; }
         public Nullable<System.DateTime> CreatedDate { get; set; }
-        public Nullable<long> ModifiedBy { get; set; }
+        public Nullable<int> ModifiedBy { get; set; }
         public Nullable<System.DateTime> ModifiedDate { get; set; }
     }
 }
